Pull database helpers into their own files
