+++ lib/nucore/database/case_sensitivity_helper.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Nucore
+
+  module Database
+
+    module CaseSensitivityHelper
+
+      def insensitive_where(relation, column, value)
+        relation.where("UPPER(#{column}) = UPPER(?)", value)
+      end
+
+    end
+
+  end
+
+end
